PM-29795: Move FileManager to data module
