+++ app/src/main/kotlin/com/x8bit/bitwarden/data/platform/manager/flightrecorder/FlightRecorderWriterImpl.kt
@@ -5,9 +5,9 @@ import android.util.Log
 import com.bitwarden.annotation.OmitFromCoverage
 import com.bitwarden.core.data.manager.dispatcher.DispatcherManager
 import com.bitwarden.core.data.util.toFormattedPattern
+import com.bitwarden.data.manager.file.FileManager
 import com.x8bit.bitwarden.BuildConfig
 import com.x8bit.bitwarden.data.platform.datasource.disk.model.FlightRecorderDataSet
-import com.x8bit.bitwarden.data.vault.manager.FileManager
 import kotlinx.coroutines.withContext
 import timber.log.Timber
 import java.io.BufferedWriter
